requirements.txt
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
pydantic>=2.0.0
python-multipart>=0.0.6
openai>=1.95.1
diff-match-patch>=20241021
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain>=0.3.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.27.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
reportlab>=4.0.7
python-magic>=0.4.27
aiofiles>=23.2.1
redis>=5.0.1
google-cloud-translate>=3.15.0
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0  